confirm dialog

+++ src/app/pages/medic/medic.component.ts
@@ -22,6 +22,7 @@ import {MedicService} from "../../services/medic.service";
 import {MedicDialogService} from "./medic-dialog.service";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {switchMap} from "rxjs";
+import {ConfirmService} from "../confirm/confirm.service";
 
 @Component({
   selector: 'app-medic',
@@ -51,7 +52,8 @@ export class MedicComponent implements OnInit{
   constructor(
     private service:MedicService,
     private dialogService: MedicDialogService,
-    private _snackBar:MatSnackBar
+    private _snackBar:MatSnackBar,
+    private confirmService:ConfirmService
 
   ) {
   }
@@ -62,6 +64,8 @@ export class MedicComponent implements OnInit{
     // Onjectos Subjects
     this.service.getMedicChange().subscribe(data => this.createTable(data));
     this.service.getMessageChange().subscribe(data => this._snackBar.open(data,"INFO",{duration:2000} ));
+    this.confirmService.getConfirm().subscribe(data => this.delete(data));
+
   }
   createTable(data: Medic[]) {
     this.dataSource = new MatTableDataSource<Medic>(data);
@@ -78,12 +82,16 @@ export class MedicComponent implements OnInit{
   openDialog(entity?:Medic){
     this.dialogService.openDialog(entity);
   }
-  delete(idPatient: number){
+  delete(idPatient: any){
+    if(!idPatient) return;
     this.service.delete(idPatient)
       .pipe(switchMap( ()=> this.service.findAll() ))
       .subscribe(data => {
         this.service.setMedicChange(data);
         this.service.setMessageChange('DELETED!');
       })
+  }
+  confirm(id:number){
+    this.confirmService.openDialog('0ms', '0ms', id);
   }
 }
